counts the chances
Server ConVars:
ttt_countablechance_det (whether a Detective can buy it - Default is 1)
ttt_countablechance_tr (whether a Traitor can buy it - Default is 1)

The probability depends on the maximum number of players that can play on the server and the number of living players. As Detective the maximum number of players is multiplied with the traitor percentage.

If you not bought it, it will count your Kills and will add them when you buy it.

If you respawn through the usage of this Item, It will count and will show as "ThirdChance", "FourthChance", ...